ko: Use correct names for month and hour in Korean
Currently, "월" and "시" are used for "month" and "hour" in Korean. This
is incorrect in this particular usage, because these are only used when
speaking about concrete months and hours, e.g. "3월" means "March" and
"2시" means "2 o'clock".

Thus, change "월" to "달" and "시" to "시간", which are actually used when
counting months and hours.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>